Print all summary

+++ config.go
@@ -13,7 +13,7 @@ type Config struct {
 
 type TaskdefConfig struct {
 	FamilyPrefix string `yaml:"family_prefix,omitempty"`
-	KeepCount    int64  `yaml:"keep_count,omitempty"`
+	KeepCount    int    `yaml:"keep_count,omitempty"`
 }
 
 func LoadConfig(path string) (*Config, error) {
